+++ webapp/src/pages/admin/Admin.tsx
@@ -5,7 +5,7 @@ import { createStore } from "solid-js/store";
 import { Top } from "../../Top";
 import * as api from "../../api";
 import { Show, createEffect, createMemo, createResource } from "solid-js";
-import { serverConfig } from "../../config";
+import { distributionName, serverConfig } from "../../config";
 
 interface AutoArchiveSettings {
   enabled: boolean;
@@ -180,7 +180,7 @@ export function Admin() {
   const links = () => {
     let links = [<>[</>];
     if (serverConfig?.datastore === "elasticsearch") {
-      links.push(<a href="/admin/elastic">Elasticsearch</a>);
+      links.push(<a href="/admin/elastic">{distributionName()}</a>);
       links.push(<> | </>);
     }
     links.push(<a href="/admin/filters">Filters</a>);
@@ -300,7 +300,10 @@ export function Admin() {
                 <Show when={serverConfig?.datastore === "elasticsearch"}>
                   <div class="row mt-2">
                     <div class="col">
-                      Warning: Do not enable if you have Elasticsearch ILM
+                      Warning: Do not enable if you have {distributionName()}{" "}
+                      {serverConfig?.distribution === "opensearch"
+                        ? "ISM"
+                        : "ILM"}{" "}
                       policies managing your indices.
                     </div>
                   </div>